refactor: api folder structure
