+++ assets/express/example-app/src/api/auth/routes/index.ts
@@ -1,0 +1,4 @@
+import login from './login.route';
+import register from './register.route';
+
+export default [login, register];
