+++ assets/express/example-app/src/api/auth/routes/register.route.test.ts
@@ -1,4 +1,4 @@
-import route from './register';
+import route from './register.route';
 
 describe('route', () => {
   it('should be defined', () => {
