+++ assets/express/example-app/src/api/healthz/routes/index.ts
@@ -1,0 +1,4 @@
+import live from './live.route';
+import ready from './ready.route';
+
+export default [live, ready];
