+++ assets/express/example-app/src/api/hello-world/index.ts
@@ -1,3 +1,3 @@
-import helloWorld from './hello-world';
+import routes from './routes';
 
-export default [helloWorld];
+export default routes;
