+++ assets/express/example-app/src/api/hello-world/routes/hello-world.route.ts
@@ -1,4 +1,4 @@
-import { RouteDefinition } from '../utils';
+import { RouteDefinition } from '@common/api';
 
 const route: RouteDefinition = {
   operationId: 'hello-world',
