+++ assets/express/example-app/src/api/hello-world/routes/index.ts
@@ -1,0 +1,3 @@
+import helloWorld from './hello-world.route';
+
+export default [helloWorld];
