+++ assets/express/example-app/src/api/v1/index.ts
@@ -1,4 +1,4 @@
+import { prefixRoutes } from '@common/api';
 import todos from './todos';
-import { prefixRoutes } from '../utils';
 
 export default prefixRoutes('/v1', [...todos]);
