+++ assets/express/example-app/src/api/v1/todos/index.ts
@@ -1,9 +1,4 @@
-import { prefixRoutes } from '../../utils';
+import { prefixRoutes } from '@common/api';
+import routes from './routes';
 
-import create from './routes/create';
-import delete_ from './routes/delete';
-import get from './routes/get';
-import list from './routes/list';
-import update from './routes/update';
-
-export default prefixRoutes('/todos', [create, delete_, get, list, update]);
+export default prefixRoutes('/todos', routes);
