+++ assets/express/example-app/src/api/v1/todos/routes/create.route.test.ts
@@ -1,4 +1,4 @@
-import route from './create';
+import route from './create.route';
 
 describe('route', () => {
   it('should be defined', () => {
