+++ assets/express/example-app/src/api/v1/todos/routes/delete.route.ts
@@ -1,4 +1,4 @@
-import { asyncHandler, defineRoute, response } from '../../../utils';
+import { asyncHandler, defineRoute, response } from '@common/api';
 import { todoIdDTO } from '../dto';
 
 export default defineRoute({
