+++ assets/express/example-app/src/api/v1/todos/routes/index.ts
@@ -1,0 +1,7 @@
+import create from './create.route';
+import delete_ from './delete.route';
+import get from './get.route';
+import list from './list.route';
+import update from './update.route';
+
+export default [create, delete_, get, list, update];
