+++ assets/express/example-app/src/api/v1/todos/routes/list.route.test.ts
@@ -1,4 +1,4 @@
-import route from './list';
+import route from './list.route';
 
 describe('route', () => {
   it('should be defined', () => {
