+++ assets/express/example-app/src/common/api/handler.ts
