+++ assets/express/example-app/src/common/api/index.ts
@@ -1,3 +1,4 @@
+export * from './middleware';
 export * from './handler';
 export * from './route';
 export * from './response';
