+++ assets/express/example-app/src/common/api/middleware/index.ts
